Move 1846 draft to step mechanism
