+++ lib/engine/round/auction.rb
@@ -59,8 +59,10 @@ module Engine
 
       def process_action(action)
         clear_cache!
+
         step = @steps.find { |s| s.actions(action.entity).include?(action.type) }
         raise GameError, "No step found for this action: #{action.to_h}" unless step
+        raise GameError, "It is not #{action.entity.name}'s turn" unless step.active_entities.include?(action.entity)
 
         step.send("process_#{action.type}", action)
       end
@@ -75,6 +77,10 @@ module Engine
 
       def next_index!
         @index = (@index + 1) % @entities.size
+      end
+
+      def reset_index!
+        @index = 0
       end
 
       private
