+++ lib/engine/step/end_game.rb
@@ -11,6 +11,10 @@ module Engine
         ACTIONS
       end
 
+      def active_entities
+        entities
+      end
+
       def process_end_game(_action)
         @game.end_game!
       end
